+++ src/Application.php
@@ -15,7 +15,6 @@ namespace WordPlate;
 
 use Dotenv\Dotenv;
 use Dotenv\Exception\InvalidPathException;
-use Symfony\Component\HttpFoundation\Request;
 
 class Application
 {
@@ -74,8 +73,9 @@ class Application
         define('NONCE_SALT', env('NONCE_SALT'));
 
         // Set the home url to the current domain.
-        $request = Request::createFromGlobals();
-        define('WP_HOME', env('WP_URL', $request->getSchemeAndHttpHost()));
+        // TODO: Remove WP_URL in the next major version.
+        $url = env('WP_URL', Request::getSchemeAndHttpHost());
+        define('WP_HOME', env('WP_HOME', $url));
 
         // Set the WordPress directory path.
         define('WP_SITEURL', env('WP_SITEURL', sprintf('%s/%s', WP_HOME, env('WP_DIR', 'wordpress'))));